Move utilities from generic to root folder
