+++ test/hasura/generic/integration-tests/data-integrity/line-routes/priority/update-line.spec.ts
@@ -1,11 +1,11 @@
 import * as config from '@config';
 import { lines } from '@datasets-generic/defaultSetup/lines';
 import { routes } from '@datasets-generic/defaultSetup/routes';
-import { getPropNameArray, queryTable, setupDb } from '@datasets-generic/setup';
 import { Line, lineProps } from '@datasets-generic/types';
 import * as dataset from '@util/dataset';
 import '@util/matchers';
 import { expectErrorResponse } from '@util/response';
+import { getPropNameArray, queryTable, setupDb } from '@util/setup';
 import * as pg from 'pg';
 import * as rp from 'request-promise';
