+++ test/hasura/generic/integration-tests/data-integrity/line-routes/validity-period/insert-route.spec.ts
@@ -2,11 +2,11 @@ import * as config from '@config';
 import { lines } from '@datasets-generic/defaultSetup/lines';
 import { routes } from '@datasets-generic/defaultSetup/routes';
 import { buildRoute } from '@datasets-generic/factories';
-import { getPropNameArray, queryTable, setupDb } from '@datasets-generic/setup';
 import { Route, RouteDirection, routeProps } from '@datasets-generic/types';
 import * as dataset from '@util/dataset';
 import '@util/matchers';
 import { expectErrorResponse } from '@util/response';
+import { getPropNameArray, queryTable, setupDb } from '@util/setup';
 import { LocalDate } from 'local-date';
 import * as pg from 'pg';
 import * as rp from 'request-promise';
