+++ test/hasura/generic/integration-tests/data-integrity/line-routes/validity-period/update-route.spec.ts
@@ -1,12 +1,12 @@
 import * as config from '@config';
 import { lines } from '@datasets-generic/defaultSetup/lines';
 import { routes } from '@datasets-generic/defaultSetup/routes';
-import { getPropNameArray, queryTable, setupDb } from '@datasets-generic/setup';
 import { Route, routeProps } from '@datasets-generic/types';
 import * as dataset from '@util/dataset';
 import { newLocalDate } from '@util/helpers';
 import '@util/matchers';
 import { expectErrorResponse } from '@util/response';
+import { getPropNameArray, queryTable, setupDb } from '@util/setup';
 import { LocalDate } from 'local-date';
 import * as pg from 'pg';
 import * as rp from 'request-promise';
