+++ test/hasura/generic/integration-tests/data-integrity/route-journey-pattern-refs/delete-infra-link-from-route.spec.ts
@@ -1,13 +1,13 @@
 import * as config from '@config';
 import { routesAndJourneyPatternsTableConfig } from '@datasets-generic/routesAndJourneyPatterns';
 import { infrastructureLinkAlongRoute } from '@datasets-generic/routesAndJourneyPatterns/routes';
-import { getPropNameArray, queryTable, setupDb } from '@datasets-generic/setup';
 import {
   InfrastructureLinkAlongRoute,
   infrastructureLinkAlongRouteProps,
 } from '@datasets-generic/types';
 import '@util/matchers';
 import { expectErrorResponse } from '@util/response';
+import { getPropNameArray, queryTable, setupDb } from '@util/setup';
 import * as pg from 'pg';
 import * as rp from 'request-promise';
