+++ test/hasura/generic/integration-tests/data-integrity/route-link-infra-link-direction/update-route-link.spec.ts
@@ -1,7 +1,6 @@
 import * as config from '@config';
 import { infrastructureLinks } from '@datasets-generic/defaultSetup/infrastructure-links';
 import { infrastructureLinkAlongRoute } from '@datasets-generic/defaultSetup/routes';
-import { getPropNameArray, queryTable, setupDb } from '@datasets-generic/setup';
 import {
   InfrastructureLinkAlongRoute,
   infrastructureLinkAlongRouteProps,
@@ -9,6 +8,7 @@ import {
 import * as dataset from '@util/dataset';
 import '@util/matchers';
 import { expectErrorResponse } from '@util/response';
+import { getPropNameArray, queryTable, setupDb } from '@util/setup';
 import * as pg from 'pg';
 import * as rp from 'request-promise';
