+++ test/hasura/generic/integration-tests/data-integrity/stop-point-infra-link-direction/insert-stop-point.spec.ts
@@ -2,12 +2,6 @@ import * as config from '@config';
 import { infrastructureLinks } from '@datasets-generic/defaultSetup/infrastructure-links';
 import { scheduledStopPoints } from '@datasets-generic/defaultSetup/scheduled-stop-points';
 import {
-  getPropNameArray,
-  getTableConfigArray,
-  queryTable,
-  setupDb,
-} from '@datasets-generic/setup';
-import {
   LinkDirection,
   ScheduledStopPoint,
   scheduledStopPointProps,
@@ -17,6 +11,12 @@ import * as dataset from '@util/dataset';
 import { serializeMatcherInputs } from '@util/dataset';
 import '@util/matchers';
 import { expectErrorResponse } from '@util/response';
+import {
+  getPropNameArray,
+  getTableConfigArray,
+  queryTable,
+  setupDb,
+} from '@util/setup';
 import { GeometryObject } from 'geojson';
 import { LocalDate } from 'local-date';
 import * as pg from 'pg';
