+++ test/hasura/generic/integration-tests/data-integrity/stop-point-infra-link-vehicle-mode/delete-infrastructure-link.spec.ts
@@ -3,11 +3,11 @@ import {
   infrastructureLinks,
   vehicleSubmodeOnInfrastructureLink,
 } from '@datasets-generic/defaultSetup/infrastructure-links';
-import { getPropNameArray, queryTable, setupDb } from '@datasets-generic/setup';
 import { infrastructureLinkProps } from '@datasets-generic/types';
 import * as dataset from '@util/dataset';
 import '@util/matchers';
 import { expectErrorResponse } from '@util/response';
+import { getPropNameArray, queryTable, setupDb } from '@util/setup';
 import * as pg from 'pg';
 import * as rp from 'request-promise';
