+++ test/hasura/generic/integration-tests/mutations/route/update.spec.ts
@@ -1,10 +1,10 @@
 import * as config from '@config';
 import { routes } from '@datasets-generic/defaultSetup/routes';
 import { buildLocalizedString } from '@datasets-generic/factories';
-import { getPropNameArray, queryTable, setupDb } from '@datasets-generic/setup';
 import { Route, routeProps } from '@datasets-generic/types';
 import * as dataset from '@util/dataset';
 import '@util/matchers';
+import { getPropNameArray, queryTable, setupDb } from '@util/setup';
 import { LocalDate } from 'local-date';
 import * as pg from 'pg';
 import * as rp from 'request-promise';
