+++ test/hasura/generic/integration-tests/mutations/scheduled-stop-point/insert-without-id.spec.ts
@@ -4,7 +4,6 @@ import {
   scheduledStopPointInvariants,
   scheduledStopPoints,
 } from '@datasets-generic/defaultSetup/scheduled-stop-points';
-import { getPropNameArray, queryTable, setupDb } from '@datasets-generic/setup';
 import {
   LinkDirection,
   ScheduledStopPoint,
@@ -13,6 +12,7 @@ import {
 } from '@datasets-generic/types';
 import * as dataset from '@util/dataset';
 import '@util/matchers';
+import { getPropNameArray, queryTable, setupDb } from '@util/setup';
 import { GeometryObject } from 'geojson';
 import { LocalDate } from 'local-date';
 import * as pg from 'pg';
