+++ test/hasura/generic/integration-tests/priority-and-validity-period-constraints/line/insert-with-conflicts.spec.ts
@@ -1,11 +1,11 @@
 import * as config from '@config';
 import { lines } from '@datasets-generic/defaultSetup/lines';
 import { buildLine, buildLocalizedString } from '@datasets-generic/factories';
-import { getPropNameArray, queryTable, setupDb } from '@datasets-generic/setup';
 import { Line, lineProps, VehicleMode } from '@datasets-generic/types';
 import * as dataset from '@util/dataset';
 import '@util/matchers';
 import { expectErrorResponse } from '@util/response';
+import { getPropNameArray, queryTable, setupDb } from '@util/setup';
 import { LocalDate } from 'local-date';
 import * as pg from 'pg';
 import * as rp from 'request-promise';
