+++ test/hasura/generic/integration-tests/priority-and-validity-period-constraints/route/insert-without-conflicts.spec.ts
@@ -2,10 +2,10 @@ import * as config from '@config';
 import { lines } from '@datasets-generic/defaultSetup/lines';
 import { routes } from '@datasets-generic/defaultSetup/routes';
 import { buildRoute } from '@datasets-generic/factories';
-import { getPropNameArray, queryTable, setupDb } from '@datasets-generic/setup';
 import { Route, RouteDirection, routeProps } from '@datasets-generic/types';
 import * as dataset from '@util/dataset';
 import '@util/matchers';
+import { getPropNameArray, queryTable, setupDb } from '@util/setup';
 import { LocalDate } from 'local-date';
 import * as pg from 'pg';
 import * as rp from 'request-promise';
