+++ test/hasura/generic/unit-tests/function-maximum_priority_validity_spans/route.spec.ts
@@ -1,10 +1,10 @@
 import * as config from '@config';
 import { buildLocalizedString } from '@datasets-generic/factories';
-import { setupDb } from '@datasets-generic/setup';
 import { Route, RouteDirection, routeProps } from '@datasets-generic/types';
 import * as db from '@util/db';
 import { nextDay, prevDay } from '@util/helpers';
 import '@util/matchers';
+import { setupDb } from '@util/setup';
 import { randomUUID } from 'crypto';
 import { LocalDate } from 'local-date';
 import * as pg from 'pg';
