+++ test/hasura/generic/unit-tests/function-maximum_priority_validity_spans/scheduled-stop-point.spec.ts
@@ -1,6 +1,5 @@
 import { networkDbConfig } from '@config';
 import { buildLocalizedString } from '@datasets-generic/factories';
-import { setupDb } from '@datasets-generic/setup';
 import {
   InfrastructureLink,
   infrastructureLinkProps,
@@ -17,6 +16,7 @@ import {
 import { singleQuery } from '@util/db';
 import { nextDay, prevDay } from '@util/helpers';
 import '@util/matchers';
+import { setupDb } from '@util/setup';
 import { randomUUID } from 'crypto';
 import { LocalDate } from 'local-date';
 import { Pool } from 'pg';
