+++ test/hasura/tsconfig.json
@@ -19,7 +19,7 @@
     "baseUrl": ".",
     "paths": {
       "@config": ["config.ts"],
-      "@util/*": ["generic/util/*"],
+      "@util/*": ["util/*"],
       "@datasets-generic/*": ["generic/datasets/*"],
       "@datasets-hsl/*": ["hsl/datasets/*"]
     }
