+++ test/hasura/util/promise.ts
